Move time controls to header
