+++ docs/less/header.less
@@ -8,7 +8,22 @@
   box-shadow: 0 2px 4px rgba(0, 0, 0, 0.5);
 }
 
-.user-info,
-.player-info {
+.user-info {
   margin-right: 1rem;
+}
+
+.time-controls {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+
+  .time-display {
+    min-width: 50px;
+    text-align: center;
+  }
+
+  button {
+    width: 40px;
+    height: 40px;
+  }
 }
