+++ docs/less/overview.less
@@ -63,23 +63,4 @@
     pointer-events: none;
   }
 
-  .time-controls {
-    position: absolute;
-    bottom: 8px;
-    right: 8px;
-    display: flex;
-    align-items: center;
-    gap: 8px;
-    color: #fff;
-
-    .time-display {
-      min-width: 50px;
-      text-align: center;
-    }
-
-    button {
-      width: 40px;
-      height: 40px;
-    }
-  }
 }
